refactor: bot.py is shared between nlp bot and watcher
