+++ packages/telegram_bots/watcher/healthcheck.py
@@ -1,6 +1,6 @@
 import requests
 from telegram_bots.config import BACKEND_URL, DEV_CHAT_ID, NLP_BOT_URL, TELEGRAM_NEXT_CHECK_TIME
-from telegram_bots.watcher.bot import watcherbot, send_message
+from telegram_bots.bot import watcherbot, send_message
 from telegram_bots import logger
 import time
